# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindEase Mental Health API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MoodEntry {
  id: number;
  userId: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  mood: number;
  note?: string | null;
  createdAt: string;
}

export interface CreateMoodEntryBody {
  /**
   * @minimum 1
   * @maximum 10
   */
  mood: number;
  note?: string;
}

export type MoodStatsTrend =
  (typeof MoodStatsTrend)[keyof typeof MoodStatsTrend];

export const MoodStatsTrend = {
  improving: "improving",
  declining: "declining",
  stable: "stable",
} as const;

export interface MoodStats {
  averageMood: number;
  totalEntries: number;
  streak: number;
  trend: MoodStatsTrend;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  reminder: "reminder",
  tip: "tip",
  achievement: "achievement",
  system: "system",
} as const;

export interface Notification {
  id: number;
  userId: string;
  title: string;
  body: string;
  type: NotificationType;
  read: boolean;
  createdAt: string;
}

export interface Conversation {
  id: number;
  userId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateConversationBody {
  title: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export type GetMoodEntriesParams = {
  days?: number;
};

export type MarkAllNotificationsRead200 = {
  count: number;
};