# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MindEase Mental Health API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: mood
    description: Mood tracking
  - name: notifications
    description: Notifications panel
  - name: openai
    description: AI chat via OpenAI
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /mood:
    get:
      operationId: getMoodEntries
      tags: [mood]
      summary: Get mood entries for current user
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of mood entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodEntry"
    post:
      operationId: createMoodEntry
      tags: [mood]
      summary: Create a mood entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoodEntryBody"
      responses:
        "201":
          description: Mood entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntry"

  /mood/stats:
    get:
      operationId: getMoodStats
      tags: [mood]
      summary: Get mood statistics summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Mood statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodStats"

  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get all notifications for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required:
                  - count

  /openai/conversations:
    get:
      operationId: getConversations
      tags: [openai]
      summary: List user conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [openai]
      summary: Create new conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /openai/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [openai]
      summary: Get messages for a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send message and stream AI response
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of AI response
          content:
            text/event-stream:
              schema:
                type: string

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MoodEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        mood:
          type: integer
          minimum: 1
          maximum: 10
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - mood
        - createdAt

    CreateMoodEntryBody:
      type: object
      properties:
        mood:
          type: integer
          minimum: 1
          maximum: 10
        note:
          type: string
      required:
        - mood

    MoodStats:
      type: object
      properties:
        averageMood:
          type: number
        totalEntries:
          type: integer
        streak:
          type: integer
        trend:
          type: string
          enum: [improving, declining, stable]
      required:
        - averageMood
        - totalEntries
        - streak
        - trend

    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [reminder, tip, achievement, system]
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - body
        - type
        - read
        - createdAt

    Conversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - createdAt
        - updatedAt

    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content