# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindEase Mental Health API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get mood entries for current user
 */
export const getMoodEntriesQueryDaysDefault = 30;

export const GetMoodEntriesQueryParams = zod.object({
  days: zod.coerce.number().default(getMoodEntriesQueryDaysDefault),
});

export const getMoodEntriesResponseMoodMax = 10;

export const GetMoodEntriesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  mood: zod.number().min(1).max(getMoodEntriesResponseMoodMax),
  note: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetMoodEntriesResponse = zod.array(GetMoodEntriesResponseItem);

/**
 * @summary Create a mood entry
 */
export const createMoodEntryBodyMoodMax = 10;

export const CreateMoodEntryBody = zod.object({
  mood: zod.number().min(1).max(createMoodEntryBodyMoodMax),
  note: zod.string().optional(),
});

/**
 * @summary Get mood statistics summary
 */
export const GetMoodStatsResponse = zod.object({
  averageMood: zod.number(),
  totalEntries: zod.number(),
  streak: zod.number(),
  trend: zod.enum(["improving", "declining", "stable"]),
});

/**
 * @summary Get all notifications for user
 */
export const GetNotificationsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  body: zod.string(),
  type: zod.enum(["reminder", "tip", "achievement", "system"]),
  read: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetNotificationsResponse = zod.array(GetNotificationsResponseItem);

/**
 * @summary Mark a notification as read
 */
export const MarkNotificationReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkNotificationReadResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  body: zod.string(),
  type: zod.enum(["reminder", "tip", "achievement", "system"]),
  read: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Mark all notifications as read
 */
export const MarkAllNotificationsReadResponse = zod.object({
  count: zod.number(),
});

/**
 * @summary List user conversations
 */
export const GetConversationsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetConversationsResponse = zod.array(GetConversationsResponseItem);

/**
 * @summary Create new conversation
 */
export const CreateConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get messages for a conversation
 */
export const GetConversationMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const GetConversationMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetConversationMessagesResponse = zod.array(
  GetConversationMessagesResponseItem,
);

/**
 * @summary Send message and stream AI response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});