# lib/api-zod/src/generated/types/createConversationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindEase Mental Health API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateConversationBody {
  title: string;
}