# lib/api-zod/src/generated/types/createMoodEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindEase Mental Health API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMoodEntryBody {
  /**
   * @minimum 1
   * @maximum 10
   */
  mood: number;
  note?: string;
}