# lib/api-zod/src/generated/types/moodEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindEase Mental Health API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MoodEntry {
  id: number;
  userId: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  mood: number;
  note?: string | null;
  createdAt: Date;
}