# lib/api-zod/src/generated/types/moodStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindEase Mental Health API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MoodStatsTrend } from "./moodStatsTrend";

export interface MoodStats {
  averageMood: number;
  totalEntries: number;
  streak: number;
  trend: MoodStatsTrend;
}